[refactor] 상세정보 SubView 분리
